refactor(Client): 第5期 — 大组件拆分与数据真实化
- FanCurveEditor 734 行拆分: 全部逻辑(拖拽/坐标映射/右键菜单/编辑弹窗/
  自动保存/服务开关/生命周期)抽为 composables/useFanCurveEditor.ts, 组件只留模板+样式
- CPU 页 4 份同构配置文件卡片改为数据驱动渲染
- 曲线生成统一到 utils/chart.ts (GPU/RyzenSmu 共用 buildSparkline)
- Home 页: 性能模式枚举↔中文映射修正为后端语义(0高性能/1静音/2平衡);
  CPU 占用改真实遥测, 噪音改为基于风扇转速的估算值并标注, 移除随机假数据
- 硬编码紫色 #8a2be2 统一为 --color-accent-purple / cyber-purple 主题类
